requirements.txt
Django
Psycopg2
djangorestframework-gis
djangorestframework-simplejwt
drf-yasg
django-countries
django-filter